fix: add alert code splitting
1. here we have done savealertjson and prepareandsavealert function to utils almost reduced 200 lines
