+++ web/src/components/alerts/AddAlert.vue
@@ -469,7 +469,6 @@ import segment from "../../services/segment_analytics";
 import {
   getUUID,
   getTimezoneOffset,
-  b64EncodeUnicode,
   b64DecodeUnicode,
   isValidResourceName,
   getTimezonesByOffset,
@@ -480,7 +479,6 @@ import useStreams from "@/composables/useStreams";
 import { outlinedInfo } from "@quasar/extras/material-icons-outlined";
 import useFunctions from "@/composables/useFunctions";
 import useQuery from "@/composables/useQuery";
-import searchService from "@/services/search";
 import { convertDateToTimestamp } from "@/utils/date";
 import {
   generateSqlQuery,
@@ -488,11 +486,15 @@ import {
 import {
   validateInputs as validateInputsUtil,
   validateSqlQuery as validateSqlQueryUtil,
+  saveAlertJson as saveAlertJsonUtil,
   type ValidationContext,
+  type JsonValidationContext,
 } from "@/utils/alertValidation";
 import {
   getAlertPayload as getAlertPayloadUtil,
+  prepareAndSaveAlert as prepareAndSaveAlertUtil,
   type PayloadContext,
+  type SaveAlertContext,
 } from "@/utils/alertPayload";
 import {
   getParser as getParserUtil,
@@ -502,7 +504,6 @@ import {
 import SelectFolderDropDown from "../common/sidebar/SelectFolderDropDown.vue";
 import AlertsContainer from "./AlertsContainer.vue";
 import JsonEditor from "../common/JsonEditor.vue";
-import { validateAlert } from "@/utils/validateAlerts";
 import { useReo } from "@/services/reodotdev_analytics";
 import {
   updateGroup as updateGroupUtil,
@@ -1055,159 +1056,36 @@ export default defineComponent({
     };
 
     const saveAlertJson = async (json: any) => {
-      let jsonPayload = JSON.parse(json);
-      let destinationsList = [];
-      props.destinations.forEach((destination: any) => {
-        destinationsList.push(destination.name);
-      });
-      let streamList = [];
-      if(!streams.value[jsonPayload.stream_type]){
-        try{
-          const response: any = await getStreams(jsonPayload.stream_type,false);
-          streams.value[jsonPayload.stream_type] = response.list;
-        }catch(err: any){
-          if(err.response.status !== 403){
-            q.notify({
-              type: "negative",
-              message: err.response?.data?.message || "Error fetching streams",
-              timeout: 3000,
-            });
-          }
-        }
-      }
-      streams.value[jsonPayload.stream_type].forEach((stream: any) => {
-        streamList.push(stream.name);
-      });
+      const saveContext: SaveAlertContext = {
+        q,
+        store,
+        props,
+        emit,
+        router,
+        isAggregationEnabled,
+        activeFolderId: { value: Array.isArray(activeFolderId.value) ? activeFolderId.value[0] : activeFolderId.value },
+        handleAlertError,
+      };
 
-      const validationResult = validateAlert(jsonPayload,{
-        streamList: streamList,
-        selectedOrgId: store.state.selectedOrganization.identifier,
-        destinationsList: destinationsList
-      });
-      
-      if (!validationResult.isValid) {
-        validationErrors.value = validationResult.errors;
-        return;
-      }
-      
-      // Validate SQL query if type is sql and not real-time
-      if (jsonPayload.is_real_time === false && jsonPayload.query_condition.type === 'sql') {
-        try {
-          // First check if query has SELECT *
-          if (!getParser(jsonPayload.query_condition.sql)) {
-            validationErrors.value = ['Selecting all columns is not allowed. Please specify the columns explicitly'];
-            return;
-          }
+      const prepareAndSaveAlertFunction = (data: any) => prepareAndSaveAlertUtil(data, saveContext);
 
-          // Set up query for validation
-          const query = buildQueryPayload({
-            sqlMode: true,
-            streamName: jsonPayload.stream_name,
-          });
+      const jsonValidationContext: JsonValidationContext = {
+        q,
+        store,
+        streams,
+        getParser,
+        buildQueryPayload,
+        prepareAndSaveAlert: prepareAndSaveAlertFunction,
+      };
 
-          delete query.aggs;
-          query.query.start_time = query.query.start_time + 780000000;
-          query.query.sql = jsonPayload.query_condition.sql;
-
-          if (jsonPayload.query_condition.vrl_function) {
-            query.query.query_fn = b64EncodeUnicode(jsonPayload.query_condition.vrl_function);
-          }
-
-          // Validate SQL query
-          await searchService.search({
-            org_identifier: store.state.selectedOrganization.identifier,
-            query,
-            page_type: jsonPayload.stream_type,
-          });
-
-          // If we get here, SQL is valid
-          showJsonEditorDialog.value = false;
-          formData.value = jsonPayload;
-          await prepareAndSaveAlert(jsonPayload);
-
-        } catch (err: any) {
-          // Handle SQL validation errors
-          const errorMessage = err.response?.data?.message || 'Invalid SQL Query';
-          validationErrors.value = [`SQL validation error: ${errorMessage}`];
-          return;
-        }
-      } else {
-        // If not SQL or is real-time, just close and update
-        showJsonEditorDialog.value = false;
-        formData.value = jsonPayload;
-        await prepareAndSaveAlert(jsonPayload);
-      }
-    };
-    const prepareAndSaveAlert = async (data: any) => {
-      const payload = cloneDeep(data);
-      if(!isAggregationEnabled.value){
-        payload.query_condition.aggregation = null;
-      }
-      if(Array.isArray(payload.context_attributes) && payload.context_attributes.length == 0){
-        payload.context_attributes = {};
-      }
-      
-      // Transform conditions to backend format
-      payload.query_condition.conditions = transformFEToBE(payload.query_condition.conditions);
-      
-      // Convert string boolean to actual boolean
-      payload.is_real_time = payload.is_real_time === "true" || payload.is_real_time === true;
-      
-      // Handle VRL function encoding if present
-      if (payload.query_condition.vrl_function) {
-        payload.query_condition.vrl_function = b64EncodeUnicode(
-          payload.query_condition.vrl_function.trim()
-        );
-      }
-
-      // Set timestamps and metadata
-      if (props.isUpdated) {
-        payload.updatedAt = new Date().toISOString();
-        payload.lastEditedBy = store.state.userInfo.email;
-        payload.folder_id = router.currentRoute.value.query.folder || "default";
-      } else {
-        payload.createdAt = new Date().toISOString();
-        payload.owner = store.state.userInfo.email;
-        payload.lastTriggeredAt = new Date().getTime();
-        payload.lastEditedBy = store.state.userInfo.email;
-        payload.folder_id = activeFolderId.value;
-      }
-
-      try {
-        const dismiss = q.notify({
-          spinner: true,
-          message: "Please wait...",
-          timeout: 2000,
-        });
-
-
-        if (props.isUpdated) {
-          await alertsService.update_by_alert_id(
-            store.state.selectedOrganization.identifier,
-            payload,
-            activeFolderId.value
-          );
-          emit("update:list", activeFolderId.value);
-          q.notify({
-            type: "positive",
-            message: "Alert updated successfully.",
-          });
-        } else {
-          await alertsService.create_by_alert_id(
-            store.state.selectedOrganization.identifier,
-            payload,
-            activeFolderId.value
-          );
-          emit("update:list", activeFolderId.value);
-          q.notify({
-            type: "positive",
-            message: "Alert saved successfully.",
-          });
-        }
-        dismiss();
-      } catch (err: any) {
-        handleAlertError(err);
-      }
+      await saveAlertJsonUtil(
+        json,
+        props,
+        validationErrors,
+        showJsonEditorDialog,
+        formData,
+        jsonValidationContext,
+      );
     };
